Refactor: Remove wrapper script and Chrome sandbox workaround
This commit refactors the project to eliminate the `run_importer.py` wrapper script.

The script was originally created as a workaround for a "Chrome Sandbox Error" that occurred when running the importer in specific environments like the Cursor editor. This approach was brittle and caused confusion by hardcoding the system Python path.

The changes in this commit are:
1. The dependency-checking logic from `run_importer.py` has been merged into the main `telegram_importer.py` script.
2. `telegram_importer.py` is now the single entry point for the application.
3. The `run_importer.py` script has been deleted.
4. The `README.md` has been updated to remove all references to `run_importer.py` and the "Chrome Sandbox Error", simplifying the usage instructions.

This makes the project cleaner, easier to understand, and more robust by removing the need for environment-specific workarounds.
