+++ README.md
@@ -17,7 +17,7 @@ Import your Telegram chat exports to WordPress with automatic formatting, catego
 
 3. **Run the importer:**
    ```bash
-   python run_importer.py
+   python telegram_importer.py
    ```
 
 ## 📋 Features
@@ -62,7 +62,6 @@ wp_import_from_tg/
 ├── wordpress_api.py          # WordPress API client
 ├── content_processor.py      # Content processing and formatting
 ├── config.py                # Configuration settings
-├── run_importer.py          # Wrapper script with dependency check
 ├── requirements.txt         # Python dependencies
 ├── env.example             # Environment variables template
 ├── categories.md           # Custom categories list
@@ -77,28 +76,28 @@ wp_import_from_tg/
 
 ### Basic Import (One by One)
 ```bash
-python run_importer.py
+python telegram_importer.py
 ```
 
 ### Batch Import
 ```bash
 # Import 10 posts starting from index 0
-python run_importer.py 0 10
+python telegram_importer.py 0 10
 
 # Continue from where you left off
-python run_importer.py 20 10
+python telegram_importer.py 20 10
 
 # Import all remaining posts
-python run_importer.py 50 0
+python telegram_importer.py 50 0
 ```
 
 ### Custom Export Directory
 ```bash
 # Use different export directory
-python run_importer.py --export-dir ChatExport_2025-07-26
+python telegram_importer.py --export-dir ChatExport_2025-07-26
 
 # Combine with batch processing
-python run_importer.py --export-dir ChatExport_2025-07-26 10 5
+python telegram_importer.py --export-dir ChatExport_2025-07-26 10 5
 ```
 
 ### Command Line Arguments
@@ -113,7 +112,7 @@ The importer shows progress and suggests the next command:
 Batch complete. Processed 10 messages, created 10 posts.
 Import completed. Next index: 20
 
-To continue, run: python run_importer.py 20 10
+To continue, run: python telegram_importer.py 20 10
 ```
 
 ## 🔧 WordPress Setup
@@ -156,28 +155,6 @@ Ensure WordPress REST API is enabled (default in modern WordPress).
 - Handles file path corrections
 
 ## 🛠️ Troubleshooting
-
-### Chrome Sandbox Error
-If you encounter this error:
-```
-The setuid sandbox is not running as root. Common causes:
-  * An unprivileged process using ptrace on it, like a debugger.
-  * A parent process set prctl(PR_SET_NO_NEW_PRIVS, ...)
-Failed to move to new namespace: PID namespaces supported, Network namespace supported, but failed: errno = Operation not permitted
-```
-
-**Solution:** This happens when running Python through Cursor (Electron-based editor). The updated `run_importer.py` automatically uses system Python to avoid this issue:
-
-```bash
-python3 run_importer.py
-```
-
-Alternatively, you can use system Python directly:
-```bash
-/usr/bin/python3 telegram_importer.py
-```
-
-**Why this happens:** Cursor is an Electron application (based on Chrome) and its embedded Python environment has Chrome dependencies that conflict with the sandbox security feature.
 
 ### Authentication Issues
 - Check WordPress REST API is enabled
